+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Added
+- Feature: Add field reason to CR Vector and VectorPipeline
+- Feature: Add ConfigCheck for Vector
+- Feature: Add ConfigCheck for VectorPipeline
 - Feature: Add utils for reconciling Kubernetes resources
 - Cleanup: Update Kustomize version in Makefile to v4.2.0 (for amd64 support)
 - Agent: Init Vector Agent Controller